requirements.txt
playwright==1.48.0
python-dotenv==1.0.1
httpx==0.27.2
schedule==1.2.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-telegram-bot==21.6
Jinja2==3.1.4
